cloud-functions/predict/requirements.txt
functions-framework==3.*
xgboost==3.2.0
scikit-learn==1.8.0
numpy>=2.0
pandas>=2.2
google-cloud-firestore>=2.16
google-cloud-bigquery>=3.14
firebase-admin>=6.4
grpcio>=1.60.0
protobuf>=4.25.0
cryptography>=41.0.0
